move syncmonitor to manager
