+++ pkg/autoscaler/autoscaler/autoscaler_manager.go
@@ -26,10 +26,8 @@ import (
 	"github.com/pingcap/tidb-operator/pkg/label"
 	"github.com/pingcap/tidb-operator/pkg/pdapi"
 	"k8s.io/apimachinery/pkg/api/errors"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/selection"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"k8s.io/apimachinery/pkg/util/sets"
 	kubeinformers "k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	appslisters "k8s.io/client-go/listers/apps/v1"
@@ -85,14 +83,6 @@ func (am *autoScalerManager) Sync(tac *v1alpha1.TidbClusterAutoScaler) error {
 	if err != nil {
 		if errors.IsNotFound(err) {
 			// Target TidbCluster Ref is deleted, empty the auto-scaling status
-			// And delete all autoscaling clusters
-			requirement, _ := labels.NewRequirement(label.AutoInstanceLabelKey, selection.Equals, []string{tac.Name})
-			err := am.cli.PingcapV1alpha1().TidbClusters(tac.Spec.Cluster.Namespace).DeleteCollection(nil, v1.ListOptions{
-				LabelSelector: labels.NewSelector().Add(*requirement).String(),
-			})
-			if err != nil {
-				klog.Errorf("cannot delete orphan autoscaling clusters for tac[%s/%s], err: %v", tac.Namespace, tac.Name, err)
-			}
 			return nil
 		}
 		return err
@@ -238,6 +228,97 @@ func (am *autoScalerManager) updateTidbClusterAutoScaler(tac *v1alpha1.TidbClust
 	})
 	if err != nil {
 		klog.Errorf("failed to update TidbClusterAutoScaler: [%s/%s], error: %v", ns, tacName, err)
+	}
+	return err
+}
+
+func (am *autoScalerManager) syncMonitor(tc *v1alpha1.TidbCluster, tac *v1alpha1.TidbClusterAutoScaler) error {
+	if monitorRef := tc.Status.Monitor; monitorRef != nil {
+		autoTcList, err := am.getAutoScaledClusters(tac, []v1alpha1.MemberType{v1alpha1.TiDBMemberType, v1alpha1.TiKVMemberType})
+		if err != nil {
+			return err
+		}
+		clusters := sets.String{}
+		clusterMap := make(map[string]*v1alpha1.TidbCluster)
+		for _, autoTc := range autoTcList {
+			fullName := fmt.Sprintf("%s/%s", autoTc.Namespace, autoTc.Name)
+			clusters.Insert(fullName)
+			clusterMap[fullName] = autoTc
+		}
+
+		monitor, err := am.tmLister.TidbMonitors(monitorRef.Namespace).Get(monitorRef.Name)
+		if err != nil {
+			return err
+		}
+
+		monitoredClusters := sets.String{}
+		newClusterRefs := make([]v1alpha1.TidbClusterRef, 0)
+		for _, tcRef := range monitor.Spec.Clusters {
+			ns := tcRef.Namespace
+			if len(ns) == 0 {
+				ns = monitor.Namespace
+			}
+
+			fullName := fmt.Sprintf("%s/%s", ns, tcRef.Name)
+			monitoredClusters.Insert(fullName)
+			tc, err := am.tcLister.TidbClusters(ns).Get(tcRef.Name)
+			if err != nil {
+				if errors.IsNotFound(err) {
+					continue
+				} else {
+					return err
+				}
+			}
+			tacName, ok := tc.Labels[label.AutoInstanceLabelKey]
+			if !ok || tacName != tac.Name {
+				// Not controlled by this AutoScaler, keep it
+				newClusterRefs = append(newClusterRefs, tcRef)
+				continue
+			}
+		}
+
+		updatedTm := monitor.DeepCopy()
+		toAdd := clusters.Difference(monitoredClusters)
+		for name := range toAdd {
+			cluster := clusterMap[name]
+			newClusterRefs = append(newClusterRefs,
+				v1alpha1.TidbClusterRef{Namespace: cluster.Namespace, Name: cluster.Name})
+		}
+		updatedTm.Spec.Clusters = newClusterRefs
+
+		err = am.updateTidbMonitor(updatedTm)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (am *autoScalerManager) updateTidbMonitor(tm *v1alpha1.TidbMonitor) error {
+	ns := tm.GetNamespace()
+	tmName := tm.GetName()
+	monitorSpec := tm.Spec.DeepCopy()
+
+	// don't wait due to limited number of clients, but backoff after the default number of steps
+	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		var updateErr error
+		_, updateErr = am.cli.PingcapV1alpha1().TidbMonitors(ns).Update(tm)
+		if updateErr == nil {
+			klog.Infof("TidbMonitor: [%s/%s] updated successfully", ns, tmName)
+			return nil
+		}
+		klog.V(4).Infof("failed to update TidbMonitor: [%s/%s], error: %v", ns, tmName, updateErr)
+		if updated, err := am.tmLister.TidbMonitors(ns).Get(tmName); err == nil {
+			// make a copy so we don't mutate the shared cache
+			tm = updated.DeepCopy()
+			tm.Spec = *monitorSpec
+		} else {
+			utilruntime.HandleError(fmt.Errorf("error getting updated TidbMonitor %s/%s from lister: %v", ns, tmName, err))
+		}
+		return updateErr
+	})
+	if err != nil {
+		klog.Errorf("failed to update TidbMonitor: [%s/%s], error: %v", ns, tmName, err)
 	}
 	return err
 }
